+++ attributes/geoip2.rb
@@ -1,9 +1,9 @@
 id = 'ngx'
 
-default[id]['geoip2']['version'] = '2.0'
+default[id]['geoip2']['version'] = '3.0'
 default[id]['geoip2']['url'] = 'https://github.com/leev/ngx_http_geoip2_module/archive'
 default[id]['geoip2']['checksum'] = \
-  'ebb4652c4f9a2e1ee31fddefc4c93ff78e651a4b2727d3453d026bccbd708d99'
+  '65a191688348a05d8d92b2e7ce9c6eb8cb8322205c34637da582a1205864133d'
 
 default[id]['geoip2']['databases']['dir'] = '/etc/chef-nginx/geolite2'
 
@@ -11,10 +11,10 @@ default[id]['geoip2']['databases']['country']['name'] = 'GeoLite2-Country.mmdb'
 default[id]['geoip2']['databases']['country']['url'] = \
   'http://geolite.maxmind.com/download/geoip/database/GeoLite2-Country.tar.gz'
 default[id]['geoip2']['databases']['country']['checksum'] = \
-  '6977a408a4f3eca76523b5bf39780fd2c8529f263ce8cc94c1da208dc9be678b'
+  '5f9565ab2ce331a1903789e0619946f23482b02fcafd8257c2f808626394b00f'
 
 default[id]['geoip2']['databases']['city']['name'] = 'GeoLite2-City.mmdb'
 default[id]['geoip2']['databases']['city']['url'] = \
   'http://geolite.maxmind.com/download/geoip/database/GeoLite2-City.tar.gz'
 default[id]['geoip2']['databases']['city']['checksum'] = \
-  '0f3c1d1115e764481962766323e838a634c16eff08e6c9229b87973ac380bb48'
+  'e367bf5084b5598e48c2eb655220945889daa8e7eff6c531e2b661bc32456701'