+++ attributes/njs.rb
@@ -1,6 +1,6 @@
 id = 'ngx'
 
-default[id]['njs']['version'] = '0.2.0'
+default[id]['njs']['version'] = '0.2.3'
 default[id]['njs']['url'] = 'https://github.com/nginx/njs/archive'
 default[id]['njs']['checksum'] = \
-  '18b12bf7b5ff00a287a9d0b802b19b8c55f2e5b1e0574cc8183d9a2648647108'
+  '4ca46da4bd4df3c1037e1dbcc95d97389db203d5b0112344a21b7ad946d16a15'