+++ attributes/ssl.rb
@@ -1,7 +1,7 @@
 id = 'ngx'
 
-default[id]['openssl']['version'] = '1.1.0h'
+default[id]['openssl']['version'] = '1.1.0i'
 default[id]['openssl']['checksum'] = \
-  '5835626cde9e99656585fc7aaa2302a73a7e1340bf8c14fd635a62c66802a517'
+  'ebbfc844a8c8cc0ea5dc10b86c9ce97f401837f3fa08c17b2cdadc118253cf99'
 
 default[id]['ssl_configuration'] = 'modern' # or intermediate